@@ -52,45 +51,44 @@
         "ng2-pdf-viewer": "^9.1.2",
         "ngx-infinite-scroll": "^14.0.0",
         "ngx-lottie": "^9.1.0",
-        "pdfjs-dist": "2.5.207",
+        "pdfjs-dist": "3.0.279",
         "rxjs": "7.5.7",
-        "tslib": "^2.3.1",
-        "twilio-video": "^2.18.3",
-        "util": "^0.12.4",
+        "tslib": "^2.4.1",
+        "util": "^0.12.5",
         "webpack": "^5.58.2",
         "webpack-dev-server": "^4.3.1",
         "zone.js": "~0.11.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "^14.2.3",
+        "@angular-devkit/build-angular": "^14.2.7",
         "@angular-eslint/builder": "14.1.2",
         "@angular-eslint/eslint-plugin": "^14.0.3",
-        "@angular-eslint/eslint-plugin-template": "13.5.0",
+        "@angular-eslint/eslint-plugin-template": "14.1.2",
         "@angular-eslint/schematics": "^14.1.2",
         "@angular-eslint/template-parser": "14.1.2",
-        "@angular/cli": "^14.2.3",
-        "@angular/compiler-cli": "^14.2.3",
-        "@angular/language-service": "^14.2.3",
-        "@types/jasmine": "~3.9.1",
+        "@angular/cli": "^14.2.7",
+        "@angular/compiler-cli": "^14.2.8",
+        "@angular/language-service": "^14.2.8",
+        "@types/jasmine": "~4.3.0",
         "@types/jasminewd2": "^2.0.10",
         "@types/lodash": "^4.14.169",
-        "@typescript-eslint/eslint-plugin": "5.27.1",
-        "@typescript-eslint/parser": "5.38.1",
+        "@typescript-eslint/eslint-plugin": "5.42.0",
+        "@typescript-eslint/parser": "5.42.0",
         "config": "^3.3.6",
-        "eslint": "^8.17.0",
+        "eslint": "^8.26.0",
         "eslint-config-prettier": "^8.5.0",
         "eslint-config-typescript": "^3.0.0",
-        "jasmine-core": "^4.4.0",
+        "jasmine-core": "^4.5.0",
         "jasmine-spec-reporter": "~7.0.0",
         "karma": "^6.3.4",
         "karma-chrome-launcher": "~3.1.0",
         "karma-coverage-istanbul-reporter": "3.0.3",
         "karma-jasmine": "^5.1.0",
-        "karma-jasmine-html-reporter": "^1.5.0",
+        "karma-jasmine-html-reporter": "^2.0.0",
         "karma-junit-reporter": "^2.0.1",
         "path-browserify": "^1.0.1",
         "prettier": "2.7.1",
-        "puppeteer": "^1.16.0",
+        "puppeteer": "^19.2.0",
         "typescript": "^4.3.5"
     }
 }
