@@ -1,7 +1,5 @@
 import { environment } from '../../environments/environment'
-import * as CryptoJS from 'crypto-js'
 import { trigger, transition, style, animate, stagger, keyframes, query } from '@angular/animations'
-import { HttpClient } from '@angular/common/http'
 
 export class Util {
     static inOutAnimation: [any] = [
@@ -108,20 +106,12 @@ export class Util {
             author: 'Code Crow Corp',
             title: 'Code Crow',
             description:
-                'A Web3 Network - Code Crow is a Stream-and-Earn platform for web3 gaming, software development, and DeFi analysis',
+                'A Web3 Network - Code Crow is a Stream-and-Earn platform for web3 gamers and developers',
             imageSrc: `${environment.hostUrl}/assets/images/logo_2.png`,
             imageWidth: '664',
             imageHeight: '358',
             url: environment.hostUrl
         }
-    }
-
-    static decrypt(value, key): string {
-        return CryptoJS.AES.decrypt(value, key).toString(CryptoJS.enc.Utf8)
-    }
-
-    static encrypt(value, key): string {
-        return CryptoJS.AES.encrypt(JSON.stringify({ value }), key).toString()
     }
 
     static compare(a: number | string, b: number | string, isAsc: boolean) {
