@@ -3,6 +3,6 @@ export const environment = {
     name: 'beta',
     hostUrl: 'https://codecrow.io',
     apiUrl: 'https://api.codecrow.io',
-    webSocketUrl: 'ws://api.codecrow.io',
+    webSocketUrl: 'wss://api.codecrow.io',
     serviceFeeWallet: 'EpVdo2d7wcaWBGhsTNydeEe6ZWtG45mrqnWR4WQxEvd4'
 }
